requirements.txt
# AutoFlow engine (desktop / library)
pyautogui==0.9.54
pyperclip==1.8.2
pynput==1.7.6
Pillow>=11.0.0
pdfplumber>=0.10.0
docling>=2.0.0
requests==2.31.0
pandas>=2.0.0

# Optional OCR fallback (only used if docling is unavailable):
# pytesseract==0.3.10
# pdf2image

# Backend API (Flask / Railway)
Flask==3.0.0
flask-cors==4.0.0
flask-limiter>=3.5.0
redis>=5.0.0
# PostgreSQL (psycopg2; SQLAlchemy postgresql+psycopg2://).
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.36,<3
alembic>=1.13.0,<2
python-dotenv==1.0.0
stripe==7.0.0
cryptography==41.0.7
gunicorn==21.2.0
resend
